refactor(agent-gui): collapse workspace context type mirror
